# packages/invenio-rest-api-client/invenio_rest_api_client-12.0.1.tar.gz/invenio_rest_api_client-12.0.1/docs/openapi/invenio.yaml
...
openapi: 3.1.1
info:
  title: Invenio REST API
  description: "## **Summary**\nThe following document is a reference guide for all the REST APIs that InvenioRDM exposes.\n## **Resources**\n- \U0001F4DA [Product documentation](https://inveniordm.docs.cern.ch)\n- \U0001F517 [OpenAPI GitHub repository](https://github.com/inveniosoftware/invenio-openapi)\n- \U0001F310 [Invenio project](https://inveniosoftware.org)\n- \U0001F4AC [Community support](https://inveniordm-dev.docs.cern.ch/install/troubleshoot/)"
  version: 12.0.0
  contact:
    name: InvenioRDM Documentation
    url: https://inveniordm.docs.cern.ch
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  x-copyright: KTH Royal Institute of Technology 2025
  x-repo: https://github.com/Samk13/invenio-api-toolkit
servers:
  - url: '{baseURL}'
    description: InvenioRDM instance
    variables:
      baseURL:
        default: https://inveniordm.web.cern.ch
        description: Base URL of your InvenioRDM instance
paths:
...
  /api/records/{recordId}/export/json:
    get:
      summary: Export record as JSON
      tags:
        - Export
      operationId: exportRecordAsJson
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the record
      responses:
        "200":
          $ref: '#/components/responses/Success'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema:
            type: object
    BadRequest:
      description: Bad request - Invalid input parameters
    Unauthorized:
      description: Unauthorized - Authentication required
    Forbidden:
      description: Forbidden - Insufficient permissions
    NotFound:
      description: Not found - Resource does not exist
    InternalServerError:
      description: Internal server error - Unexpected server error
...
security:
  - BearerAuth: []
...